refactor: move WikiBackground and Sidebar outof Main
