+++ src/constants/style.ts
@@ -1,0 +1,1 @@
+export const sidebarWidth = 68;
